chore: Ensure typecheck on lint, avoid global DOM types, improve test
This is a general maintenance patch. The lint script is adapted to use
the linting config, so that test files also get typechecked. By setting
"lib" in TSConfig explicitly, the default DOM types are avoided, which
would pollute the global namespace. Finally, some import behavior in the
test file is improved by using extensions via the experimental ts-node
resolver, and getting rid of a bogus mocha import.